% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntrf.R
\docType{data}
\name{doughnut}
\alias{doughnut}
\title{Doughnut shaped dataset}
\format{An object of class \code{data.frame} with 10000 rows and 3 columns.}
\usage{
data(doughnut)
}
\description{
Doughnut-shaped two-class 2-dimensinoal classification problem
V1 and V2 are the input features, normalized to 0-1
V3 is the output (TRUE / FALSE)
}
\examples{
data(doughnut)
plot(doughnut$V1, doughnut$V2, col=doughnut$V3)
}
\keyword{datasets}

