% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntrf.R
\docType{data}
\name{doughnutRand}
\alias{doughnutRand}
\title{Doughnut shaped dataset with 8 extra random attributes}
\format{An object of class \code{data.frame} with 10000 rows and 11 columns.}
\usage{
data(doughnutRand)
}
\description{
V1 to V8 are the extra random features (in the range 0-1)
V9 and V10 are the original input features, normalized to 0-1
V11 is the output (TRUE / FALSE)
}
\examples{
data(doughnutRand)
plot(doughnutRand$V9, doughnutRand$V10, col=doughnutRand$V11)
}
\keyword{datasets}

