% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntrf.R
\docType{data}
\name{doughnutRandRotated}
\alias{doughnutRandRotated}
\title{Doughnut shaped dataset with 8 extra random attributes and rotated}
\format{An object of class \code{data.frame} with 10000 rows and 11 columns.}
\usage{
data(doughnutRandRotated)
}
\description{
doughnutRandRotated randomly rotated
}
\keyword{datasets}

