% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntrf.R
\name{nntrf_doughnut}
\alias{nntrf_doughnut}
\title{{nntrf_doughnut: a benchmark of three doughnut datasets}}
\usage{
nntrf_doughnut(verbose = TRUE)
}
\arguments{
\item{verbose}{(default TRUE) Print results to the console.}
}
\value{
list of: \itemize{
  \item Accuracies of KNN on the Doughnut Dataset: with no nntrf, with nntrf and no sigmoid, and with nntrf and no sigmoid and 5 repetitions
  \item Accuracies of KNN on the DoughnutRand Dataset: with no nntrf, with nntrf and no sigmoid, and with nntrf and no sigmoid and 5 repetitions
  \item Accuracies of KNN on the DoughnutRandRotated Dataset: with no nntrf, with nntrf and no sigmoid, and with nntrf and no sigmoid and 5 repetitions
}
}
\description{
{This function compares KNN with data untransformed vs. data transformed with nntrf.
In order to see what it does, check the code: nntrf::nntrf_doughnut
}
}

