% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntrf.R
\name{nntrf_iris}
\alias{nntrf_iris}
\title{{nntrf_iris: a benchmark for the iris dataset}}
\usage{
nntrf_iris(verbose = TRUE)
}
\arguments{
\item{verbose}{(default TRUE) Print results to the console.}
}
\value{
Accuracies of KNN on the Iris Dataset: with no nntrf, and with nntrf and no sigmoid
}
\description{
{This function compares KNN with data untransformed vs. data transformed with nntrf.
In order to see what it does, check the code: nntrf::nntrf_iris
}
}

